refactor: change implementation of piece factory
each subclass now implements a #handles_notation? method that indicates
what FEN character symbolizes that piece-type. this will simplify logic
elsewhere in the application, and it also simplifies the factory method
in Piece.

it does tie things more closely to FEN, but I don't think this is an
issue because of the domain and how common these piece abbreviations
are.