feat: add basic move verification process
move lists can now be filtered according to #legal_moves and whether the
move coordinate is on the board as well as whether the move coordinate is
either empty or occupied by an enemy, meaning it's open for capture. More
sophisticated capture logic will be necessary for pieces other than the
Knight, and self-check moves will need to be considered as well.