refactor: delete Move library and spec files
This refactor's purpose was to clean up the interaction between
component's of the program for creating a piece's list of moves. The
Move class was a premature abstraction/optimization that now doesn't
fit with the flow of the rest of the project. All of this functionality
has been cleanly moved into the Pieces themselves.