feat: separate Pawn into White and Black Pawn
This felt like an essential change because of how pawns have
fundamental differences depending on color, unlike the other pieces
involved in the game. This will make managing their movements easier.
I didn't feel the need to have them inherit from a Pawn superclass as
I think shared behavior that can be abstracted to a superclass will be
fairly minimal.