refactor: begin large refactor of Board to leverage new Square class
Refactored two instantiation methods (both initialize and from_fen) to
make use of Square. Now the Board will be created with 64 square objects
that represent the cells on a chess board. These squares are also responsible
for some of the piece handling that the Board has been implementing, which
has been bloating the Board class.