feat: introduce rudimentary castling implementation
This commit contains the first castle move (king side castling for white).
The other three will follow. Currently these moves only move the King.
Further work will be done in the Chess class to move the involved Rook
if the move is a castling move. The Chess class is also where toggling
the availability for castling will occur.