refactor: Rook and King's #involved_in_castling? uses position
This change reflects the fact that these pieces are only involved
in castling if they are in their starting positions. These pieces are still
not for certain able to participate in a castling move even if they are
located in their starting positions, but other aspects of the game will
cover these edge cases.