feat: add #to_string_with_moves method
This method allows a square to be displayed in the context of a user
viewing potential moves. Squares that are in the move list, and can
be moved to, will be displayed to the user differently than other squares.
Empty squares that can be moved to are marked with a black dot. Squares
with enemy pieces will have a red background color -- indicating a capture
opportunity.